pyproject.toml
[project]
name = "end-to-end-data-engineering-and-ml-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiokafka==0.10.0",
    "asyncio==3.4.3",
    "kafka-python==2.0.2",
    "mlflow==2.9.2",
    "numpy==1.26.2",
    "pandas==2.1.4",
    "plotly==5.18.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "pyspark==3.5.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dotenv==1.0.0",
    "setuptools>=80.9.0",
    "streamlit==1.29.0",
    "structlog==24.1.0",
]
